pyproject.toml
[tool.poetry]
name = "ayashige"
version = "0.1.0"
description = ""
authors = ["Manabu Niseki <manabu.niseki@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
aiocache = {extras = ["redis"], version = "^0.12.0"}
aiofiles = "^23.1.0"
arq = "^0.25"
dataclasses-json = "^0.5.7"
fastapi = "^0.92"
gunicorn = "^20.1.0"
httpx = "^0.23"
Levenshtein = "^0.20.8"
loguru = "^0.6.0"
pyhumps = "^3.8.0"
pyopenssl = "^23.0"
redis = {extras = ["hiredis"], version = "^4.5.1"}
tld = "^0.13"
ujson = "^5.7"
uvicorn = {extras = ["standard"], version = "^0.20"}
websocket-client = "^1.5"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.0"
autopep8 = "^2.0"
black = "^23.1"
coveralls = "^3.3.1"
flake8 = "^6.0"
isort = "^5.12"
mypy = "^1.0"
nest-asyncio = "^1.5.6"
pre-commit = "^3.1"
py = "^1.11.0"
pytest = "^7.2"
pytest-asyncio = "^0.20"
pytest-cov = "^4.0.0"
pytest-env = "^0.8.1"
pytest-mock = "^3.10.0"
pytest-parallel = "^0.1.1"
pytest-randomly = "^3.12.0"
pytest-sugar = "^0.9"
pytest-timeout = "^2.1.0"
pyupgrade = "^3.3"
respx = "^0.20.0"
vcrpy = "^4.2.1"
types-redis = "^4.5.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
env = [
    "TESTING=True"
]
